+++ components/form/src/index.js
@@ -1,4 +1,4 @@
-import { Form } from './src/auro-component-form.js';
+import { AuroForm } from './src/auro-form.js';
 
 /**
  * Register Custom Element.
@@ -8,7 +8,7 @@ import { Form } from './src/auro-component-form.js';
 const registerComponent = (name = 'custom-form') => {
   // alias definition
   if (!customElements.get(name)) {
-    customElements.define(name, class extends Form {});
+    customElements.define(name, class extends AuroForm {});
   }
 };
